review: address Copilot PR feedback (6 fixes)
- HomeView (Swift demo): subscribeToEvents() leaked a detached Task.
  Switched to addEventListener as the single source of truth so the
  PayabliTTPEventToken in onDisappear cancels the actual subscription
  instead of an unused no-op listener.
- PayabliFlutterDemo/main.dart: fixed PayabliTTPSessionState references
  (.Ready → .ready, .Error → .error, .SessionExpired → .sessionExpired,
  .PendingActivation → .pendingActivation) — the demo wouldn't compile
  against the Dart enum (which is lowerCamelCase per Dart conventions).
- PayabliTTP @objc init: wrapped tokenRefreshHandler bridging in a
  one-shot Locked<Bool> guard so a host that invokes the completion
  block twice doesn't double-resume the CheckedContinuation (which
  would crash). Subsequent calls are silently ignored. Added a small
  NSLock-backed Locked<T> reference cell as an internal helper next to
  UncheckedSendableBox.
- build_release_frameworks.sh: removed the stale "PAYIN_SHA256,
  CARDREADER_SHA256" comment — the script now exposes
  CORE_SHA256 / TAPTOPAY_SHA256 / CARD_READER_CORE_SHA256.
- PayabliMAUIDemo MainPage.xaml.cs: fixed Charge() invocation —
  C# enums don't convert directly to nint/IntPtr. The binding signature
  expects nint, so cast PayabliTTPPaymentType.Sale through its long
  underlying type: (nint)(long)PayabliTTPPaymentType.Sale.

iOS xcodebuild test: PayabliSDKCoreTests + PayabliSDKTapToPayTests +
PayabliSDKTelemetryTests all pass.

+++ Example/PayabliMAUIDemo/MainPage.xaml.cs
@@ -115,7 +115,10 @@ public partial class MainPage : ContentPage
         SetWorking(true);
         _ttp.Charge(
             amount: NSDecimalNumber.FromString(amount.ToString(System.Globalization.CultureInfo.InvariantCulture)),
-            type: (System.IntPtr)PayabliTTPPaymentType.Sale,
+            // PayabliTTPPaymentType is `enum : long`, and the binding signature
+            // expects `nint`. C# disallows enum -> nint without going through
+            // the underlying integral type first, so cast through `long`.
+            type: (nint)(long)PayabliTTPPaymentType.Sale,
             serviceFee: NSDecimalNumber.Zero,
             customer: null,
             order: null,
